Refactored pop up menu for projects.
